// src/app/core/models/user.model.ts
export class User {
  id?: number;
  username: string;
  password: string;
  firstName?: string;
  lastName?: string;
  role?: Role;
  AuthToken?: string;
  clientId?: string;
  clientName?: string;
  path?: string;
  validation?: string;
  alert?: string;
  ClientId?: number;
  Organizetion?: number;
  LoginId?: string;
  UserId?: number;
  IsTempPassword?: boolean;
  SourceSystemID?: number;
}